TrickyCurve_updateEffectHandleRing: (s16) subtrahend cast drops extsh
variantSfxTimer -= (s16)(int)timeDelta carried a redundant (int); (s16)
timeDelta (#53) drops the extsh on the s16 RMW. 79.66->80.29%.

+++ src/main/dll/DF/dll_0232_dfprotatep.c
@@ -480,8 +480,7 @@ void TrickyCurve_updateEffectHandleRing(int obj)
 
     if (state->variantSfxTimer != 0 && flags->bit10 != 0)
     {
-        state->variantSfxTimer -= (s16)(int)
-        timeDelta;
+        state->variantSfxTimer -= (s16)timeDelta;
         if (state->variantSfxTimer <= 0)
         {
             state->variantSfxTimer = 200;
